completed, but has errors
There are errors in the prospector script.   I suspect that I have code that should be at the end of the file when it's in the middle.
Having issues attaching scripts to the in game text, met with error message stating that the script needs to be in monobehaviour, which the script is.